requirements.txt
# Core dependencies
torch>=2.0.0
open-clip-torch>=2.20.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Phase 2: FAISS integration
faiss-cpu>=1.7.4

# Phase 3: Explainability
matplotlib>=3.7.0

# Phase 4: Configuration
pyyaml>=6.0

# Development and testing
pytest>=7.4.0
